ISSUE-16 - Create Graph Module
Pull code for build_graph into own module, with dependency graph
structure.
