+++ cargo-geiger/src/cli.rs
@@ -14,6 +14,7 @@
 use cargo::CliError;
 use crate::format::Pattern;
 use crate::Args;
+use crate::graph::{Graph, Node};
 use cargo::core::compiler::CompileMode;
 use cargo::core::compiler::Executor;
 use cargo::core::compiler::Unit;
@@ -41,10 +42,8 @@ use geiger::Count;
 use geiger::CounterBlock;
 use geiger::IncludeTests;
 use geiger::RsFileMetrics;
-use petgraph::graph::NodeIndex;
-use petgraph::visit::EdgeRef;
 use petgraph::EdgeDirection;
-use std::collections::hash_map::Entry;
+use petgraph::visit::EdgeRef;
 use std::collections::{HashMap, HashSet};
 use std::ffi::OsString;
 use std::io;
@@ -62,13 +61,6 @@ use walkdir::WalkDir;
 pub enum Charset {
     Utf8,
     Ascii,
-}
-
-pub enum ExtraDeps {
-    All,
-    Build,
-    Dev,
-    NoMore,
 }
 
 #[derive(Clone, Copy)]
@@ -93,18 +85,6 @@ pub struct PrintConfig<'a> {
     pub charset: Charset,
     pub allow_partial_results: bool,
     pub include_tests: IncludeTests,
-}
-
-pub struct Node {
-    id: PackageId,
-    // TODO: Investigate why this was needed before the separation of printing
-    // and graph traversal and if it should be added back.
-    //pack: &'a Package,
-}
-
-pub struct Graph {
-    graph: petgraph::Graph<Node, DepKind>,
-    nodes: HashMap<PackageId, NodeIndex>,
 }
 
 /// TODO: Write proper documentation for this.
@@ -133,72 +113,6 @@ pub fn get_cfgs(
     ))
 }
 
-/// Almost unmodified compared to the original in cargo-tree, should be fairly
-/// simple to move this and the dependency graph structure out to a library.
-/// TODO: Move this to a module to begin with.
-pub fn build_graph<'a>(
-    resolve: &'a Resolve,
-    packages: &'a PackageSet,
-    root: PackageId,
-    target: Option<&str>,
-    cfgs: Option<&[Cfg]>,
-    extra_deps: ExtraDeps,
-) -> CargoResult<Graph> {
-    let mut graph = Graph {
-        graph: petgraph::Graph::new(),
-        nodes: HashMap::new(),
-    };
-    let node = Node {
-        id: root,
-        //pack: packages.get_one(root)?,
-    };
-    graph.nodes.insert(root, graph.graph.add_node(node));
-
-    let mut pending = vec![root];
-
-    while let Some(pkg_id) = pending.pop() {
-        let idx = graph.nodes[&pkg_id];
-        let pkg = packages.get_one(pkg_id)?;
-
-        for raw_dep_id in resolve.deps_not_replaced(pkg_id) {
-            let it = pkg
-                .dependencies()
-                .iter()
-                .filter(|d| d.matches_ignoring_source(raw_dep_id.0))
-                .filter(|d| extra_deps.allows(d.kind()))
-                .filter(|d| {
-                    d.platform()
-                        .and_then(|p| {
-                            target.map(|t| match cfgs {
-                                None => false,
-                                Some(cfgs) => p.matches(t, cfgs),
-                            })
-                        })
-                        .unwrap_or(true)
-                });
-            let dep_id = match resolve.replacement(raw_dep_id.0) {
-                Some(id) => id,
-                None => raw_dep_id.0,
-            };
-            for dep in it {
-                let dep_idx = match graph.nodes.entry(dep_id) {
-                    Entry::Occupied(e) => *e.get(),
-                    Entry::Vacant(e) => {
-                        pending.push(dep_id);
-                        let node = Node {
-                            id: dep_id,
-                            //pack: packages.get_one(dep_id)?,
-                        };
-                        *e.insert(graph.graph.add_node(node))
-                    }
-                };
-                graph.graph.add_edge(idx, dep_idx, dep.kind());
-            }
-        }
-    }
-
-    Ok(graph)
-}
 
 pub fn get_workspace(
     config: &Config,
@@ -589,18 +503,6 @@ fn get_kind_group_name(k: DepKind) -> Option<&'static str> {
 }
 
 // ---------- END: Public items ----------
-
-impl ExtraDeps {
-    fn allows(&self, dep: DepKind) -> bool {
-        match (self, dep) {
-            (_, DepKind::Normal) => true,
-            (ExtraDeps::All, _) => true,
-            (ExtraDeps::Build, DepKind::Build) => true,
-            (ExtraDeps::Dev, DepKind::Development) => true,
-            _ => false,
-        }
-    }
-}
 
 #[derive(Debug)]
 enum RsResolveError {
